pyproject.toml
[tool.poetry]
name = "tech-accidents"
version = "0.1.0"
description = ""
authors = ["Артем Баландин <\"tom9981@yandex.ru>"]
readme = "README.md"
packages = [{include = "tech_accidents"}]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.21.1"
sqlalchemy = "^2.0.9"
alembic = "^1.10.3"
asyncpg = "^0.27.0"
aiosqlite = "^0.19.0"
fastapi-users = {extras = ["sqlalchemy"], version = "^12.0.0"}
python-dotenv = "^1.0.0"
structlog = ">=23.1.0"
pydantic-settings = "^2.0.3"
requests = "^2.31.0"
asgi-correlation-id = "4.2.0"
pydantic = "2.6.4"
numpy = "^1.26.4"


[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
